testutils/common: artifact creator helper
useful contextmanager lifted from one of the tests

changelog: none
